Added Save Button
Right clicking an html canvas no longer works, so I added a save button to download the canvas as a png. However, this causes CORS issues when run as a plain html. This can be fixed by running it in a local web server and navigating to it in a browser.