Update README.MD
To reflect changes to the html. Instead of just clicking the file, now it has to be run in a web server to avoid CORS issues in modern browsers. An easy way is to just use the http.server that comes with python, and I added the instructions to do so.